fix : 쿼리 수정 , redis 캐시 수정

+++ backend/src/main/java/com/yata/backend/domain/member/dto/MemberDto.java
@@ -61,7 +61,7 @@ public class MemberDto {
         private Member.MemberStatus memberStatus;
         private Double fuelTank;
         private List<String> roles;
-        private Long yataCount;
+        private int yataCount;
 
     }
