+++ backend/src/main/java/com/yata/backend/domain/yata/repository/yataMemberRepo/YataMemberRepository.java
@@ -2,5 +2,5 @@ package com.yata.backend.domain.yata.repository.yataMemberRepo;
 
 
 public interface YataMemberRepository {
-    Long countByMember_email(String email);
+    Integer countByMember_email(String email);
 }
